backend - revision(select data spinner and kurang stok jika ditambah) v2

+++ app/src/main/java/com/projek/iwanmotor/ui/barang/TambahBarang.kt
@@ -107,8 +107,8 @@ class TambahBarang : Fragment() {
             viewModel.updateItem(
                 this.navigationArgs.itemId,
                 binding.inputNamaProduk.text.toString(),
-                binding.inputHargaModal.text.toString(),
-                binding.inputHargaJual.text.toString(),
+                binding.inputHargaModal.text.toString().toDouble(),
+                binding.inputHargaJual.text.toString().toDouble(),
                 binding.inputStok.text.toString(),
                 binding.inputTglMasuk.text.toString()
             )
